add import

+++ cli/commands.py
@@ -3,6 +3,7 @@ from enum import Enum
 
 import binascii
 from cli import crypto, file_io
+from cli.libs.amount import int_to_amount
 from cli.exception import CliException
 from primitive_pb2 import Amount, uint256
 from commands_pb2 import Command, CreateAsset, AddAssetQuantity, CreateAccount, CreateDomain, TransferAsset
@@ -42,7 +43,7 @@ class CommandList:
                         "required": True
                     },
                     "amount": {
-                        "type": int,
+                        "type": float,
                         "detail": "target's asset id like japan/yen",
                         "required": True
                     },
@@ -127,7 +128,7 @@ class CommandList:
                         "required": False
                     },
                     "amount": {
-                        "type": int,
+                        "type": float,
                         "detail": "how much transfer",
                         "required": True
                     }
